Update Startup.cs
You need to add en-US culture as your code in the controller expect this format. I sit here in Denmark, and the demo does not take into account a locale different from en-US. Eg. instead of 12$ the result is 120$.

Reference:

Task 4: Run the application in the development environment
In the ShirtStore - Microsoft Visual Studio window, on the toolbar, click the arrow next to the Start Debugging button, and then click Development.​

In the ShirtStore - Microsoft Visual Studio window, on the DEBUG menu, click Start Without Debugging.

In the ShirtStore - Microsoft Visual Studio window, on the Output tab, in the Show output from list, select ASP.NET Core Web Server, and then click the Clear All button.

In Microsoft Edge, on the Size list, select M.

In Microsoft Edge, on the Color list, select Yellow.

In Microsoft Edge, on the Price box, type 10.

In Microsoft Edge, on the Tax box, type 1.2.

In Microsoft Edge, click Add Shirt to Stock.